+++ src/main/java/com/example/team5_be/trophy/dao/TrophyRepository.java
@@ -9,6 +9,6 @@ import com.example.team5_be.habit.domain.entity.HabitEntity;
 import com.example.team5_be.trophy.domain.entity.TrophyEntity;
 
 @Repository
-public interface TrophyRepository extends JpaRepository<TrophyEntity, Long> {
+public interface TrophyRepository extends JpaRepository<TrophyEntity, Integer> {
     Optional<TrophyEntity> findByHabit(HabitEntity habit);
 }
